[#1606] improvement(client): Add retry mechanism for reading data(memory/local/index) when there is no buffer
